Switch host builds to use compiler-rt
Previously the host builds didn't use compiler-rt due to us lacking a
host version of libunwind.  Now that the bootstrap system builds the
in-tree libunwind and links against that the host versions can use
compiler-rt.

The device versions are made to use compiler-rt via a set of shims in
the NDK, so no changes are needed there.

Test: ./build.py --lto thin
Change-Id: I910e6a4d546504c391b3c8fe1ea36f86d52b1845