Added a sysroot argument to host compiler/linker wrappers
This CL modifies the host compiler and linker wrapper templates to add a
sysroot flag.  This revealed that the existing configuration was
non-hermetic and the build system was picking up host crt object files.
The fix for this issue was to add an additional library search path and
correct the existing binary search path arguments.

Test: ./build.py --lto thin
Change-Id: I3e546363830cf3a056b3a030a18185e730f5ec66