Add arguments for controlling PGO
This CL adds the `--profile-generate` and `--profile-use` flags to the
build.py script.  These options can be used to enable the generation
of either an instrumented or an profile-guided optimized version of
the rustc compiler.

The vp-counters-per-site option is used to change the value from the
default 1.5 to 2 due to runtime warnings about the instrumented binary
running out of counters.

Test: ./build.py --lto thin --profile-generate --build-name pgo-stage1
Test: ./update_prebuilts.py ../../dist/rust-pgo-stage1.tar.gz 1.58.1p1
Test: m rust && m
Test: llvm-profdata -o out/profiles/llvm.profdata out/profiles/llvm/*
Test: llvm-profdata -o out/profiles/rustc.profdata out/profiles/rustc/*
Test: ./build.py --lto thin --profile-use --build-name pgo-stage2
Test: ./update_prebuilts.py ../../dist/rust-pgo-stage2.tar.gz 1.58.1p2
Test: m rust && m

Bug: 218368713
Change-Id: I3b96293ea3d14a8ae511314b3eeee32f59bf1743