Fixed a bug in update_prebuilts
Previously it was possible for the script to reach a bad configuration
when the target-and-version directory existed but the overwrite flag
was passed.  This was caused by Git's behavior of removing empty
directories, even if they weren't specified as part of a `git rm`
command.

This CL modifies the script to ensure that the directory is re-created
after it is emptied as part of an overwrite operation.

Test: Run script with overwrite on a repo that has an existing branch
Change-Id: I85c64ae6e7daf9f49cf094e7d658c823369ec1e0