Re-work compiler/linker flag handling
Until recently the set of flags passed to the compilers and linkers
involved in building the Rust toolchain was both small and static.  With
the addition of new features to the build process there are now more
flags being passed and most of them are added dynamically.

With the flags split between the static files and an evolved (as opposed
to designed) bit of Python code it was sometimes hard to determine where
flags would end up.  In addition, some flags had to be duplicated to
ensure they were passed to all of the desired destinations.

This CL re-works the flag configuration process by creating a hierarchy
of flag sets that are combined in a structured fashion.  The logic for
adding flags has also been re-organized to move related changes closer
together.  The final set of flags are then printed the console, making
it easier to audit a build invocation.

The PGO staging for rustc was also adjusted to support PGO
instrumentation with statically-linked LLVM libraries.

Test: ./build.py
Test: ./build.py --lto thin
Test: ./build.py --lto thin --emit-relocs --gc-sections
--profile-generate

Change-Id: I503bc536c61260a2f27c1d6467506f7ab6b47769